Convert to TSX (WOR-1800).
